Integrate planejamento module into main application
