+++ tests/test_planejamento_api.py
@@ -3,12 +3,11 @@ from datetime import date
 
 from openpyxl import Workbook, load_workbook
 
-from app.planejamento.models import Planejamento
-from src.models import db
+from src.models import db, Planejamento
 from src.models.instrutor import Instrutor
 
 
-def test_api_planejamento_export(client):
+def test_api_planejamento_export(client, non_admin_auth_headers):
     with client.application.app_context():
         instr = Instrutor(nome="Instrutor Teste")
         db.session.add(instr)
@@ -27,15 +26,18 @@ def test_api_planejamento_export(client):
         )
         db.session.add(p)
         db.session.commit()
-    resp = client.get("/api/planejamento/export?mes=2024-01", headers={"X-Role": "ROLE_USER"})
+    resp = client.get(
+        "/planejamento/api/planejamento/export?mes=2024-01",
+        headers=non_admin_auth_headers,
+    )
     assert resp.status_code == 200
     wb = load_workbook(filename=io.BytesIO(resp.data))
     ws = wb.active
     assert ws.title == "Planejamento"
     assert ws.cell(row=2, column=1).value == "05/01/2024"
-    assert ws.cell(row=2, column=7).value == "Instrutor Teste"
+    assert ws.cell(row=2, column=4).value == "Instrutor Teste"
 
-def test_api_planejamento_import(client):
+def test_api_planejamento_import(client, login_admin):
     wb = Workbook()
     ws = wb.active
     ws.append([
@@ -66,11 +68,13 @@ def test_api_planejamento_import(client):
     wb.save(stream)
     stream.seek(0)
     data = {"file": (stream, "import.xlsx")}
+    token, _ = login_admin(client)
+    headers = {"Authorization": f"Bearer {token}"}
     resp = client.post(
-        "/api/planejamento/import",
+        "/planejamento/api/planejamento/import",
         data=data,
         content_type="multipart/form-data",
-        headers={"X-Role": "ROLE_ADMIN"},
+        headers=headers,
     )
     assert resp.status_code == 200
     assert resp.json["ok"] is True
@@ -83,7 +87,7 @@ def test_api_planejamento_import(client):
         assert instrutor.nome == "Instrutor Novo"
 
 
-def test_api_planejamento_conflict(client):
+def test_api_planejamento_conflict(client, login_admin):
     with client.application.app_context():
         instr = Instrutor(nome="Instrutor Conflito")
         db.session.add(instr)
@@ -104,20 +108,22 @@ def test_api_planejamento_conflict(client):
         db.session.add_all([p1, p2])
         db.session.commit()
         pid2 = p2.id
+    token, _ = login_admin(client)
+    headers = {"Authorization": f"Bearer {token}"}
     resp = client.post(
-        "/api/planejamento",
+        "/planejamento/api/planejamento",
         json={
             "data": "2024-01-01",
             "turno": "MANHA",
             "treinamento": "Curso X",
             "instrutor_id": instr_id,
         },
-        headers={"X-Role": "ROLE_ADMIN"},
+        headers=headers,
     )
     assert resp.status_code == 409
     resp2 = client.put(
-        f"/api/planejamento/{pid2}",
+        f"/planejamento/api/planejamento/{pid2}",
         json={"data": "2024-01-01", "turno": "MANHA"},
-        headers={"X-Role": "ROLE_ADMIN"},
+        headers=headers,
     )
     assert resp2.status_code == 409
